aritm_vr

+++ L4/main.c
@@ -91,5 +91,5 @@ double srednjaVr(){
         scanf("%lf", br[i]);
         zbir+=br[i];
     }
-    return zbir/n;
+    return (double)zbir/n;
 }
